+++ apps/emails-and-messages/scripts/migrations/migration-utils.ts
@@ -1,11 +1,14 @@
 /* eslint-disable turbo/no-undeclared-env-vars */
 
-import { createGraphQLClient } from "@saleor/apps-shared";
 import { SaleorCloudAPL } from "@saleor/app-sdk/APL";
 import { createSettingsManager } from "../../src/lib/metadata-manager";
+import { createInstrumentedGraphqlClient } from "../../src/lib/create-instrumented-graphql-client";
 
 export const getMetadataManagerForEnv = (apiUrl: string, appToken: string, appId: string) => {
-  const client = createGraphQLClient({ saleorApiUrl: apiUrl, token: appToken });
+  const client = createInstrumentedGraphqlClient({
+    saleorApiUrl: apiUrl,
+    token: appToken,
+  });
 
   return createSettingsManager(client, appId);
 };