+++ apps/emails-and-messages/src/modules/channels/channels.router.ts
@@ -1,11 +1,14 @@
 import { ChannelsFetcher } from "./channels-fetcher";
-import { createGraphQLClient } from "@saleor/apps-shared";
 import { router } from "../trpc/trpc-server";
 import { protectedClientProcedure } from "../trpc/protected-client-procedure";
+import { createInstrumentedGraphqlClient } from "../../lib/create-instrumented-graphql-client";
 
 export const channelsRouter = router({
   fetch: protectedClientProcedure.query(async ({ ctx }) => {
-    const client = createGraphQLClient({ saleorApiUrl: ctx.saleorApiUrl, token: ctx.token });
+    const client = createInstrumentedGraphqlClient({
+      saleorApiUrl: ctx.saleorApiUrl,
+      token: ctx.token,
+    });
 
     const fetcher = new ChannelsFetcher(client);