+++ apps/slack/CHANGELOG.md
@@ -1,5 +1,12 @@
 # saleor-app-slack
 
+## 1.10.1
+
+### Patch Changes
+
+- Updated dependencies [67afe8e4]
+  - @saleor/apps-shared@1.9.3
+
 ## 1.10.0
 
 ### Minor Changes