+++ packages/webhook-utils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @saleor/webhook-utils
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies [67afe8e4]
+  - @saleor/apps-shared@1.9.3
+
 ## 0.0.3
 
 ### Patch Changes